fix: 年齡欄位改為允許小數；修正LIFF編輯毛孩無法顯示已儲存資料的問題；後台新增完整編輯寵物頁面

+++ src/main/java/com/petgrooming/pet_system/dto/PetRequest.java
@@ -25,7 +25,7 @@ public class PetRequest {
 
     @NotNull(message = "年齡不能為空")
     @PositiveOrZero(message = "年齡不能為負數")
-    private Integer age;
+    private Double age; // 需求（追加）：允許小數，例如 5.5 歲，方便標記幼犬/幼貓的半歲
 
     // 需求 2：毛長不再由顧客填寫（改由店家後台定義），故此處移除 coatType 欄位。
     // 前端即使誤傳此欄位，也會被忽略（Spring Boot 預設不因未知屬性報錯）。
